Change AuthorizeRoute to React.FC
